gptSite/pyproject.toml
[project]
name = "gptsite"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
]
